Re-phrase usage envs in typing rules
Previously, the usage environment attached to a variable was said to
contain all variables needed to typecheck the expression bound to that
variable. However, this made the substitution lemma proof impossible, so
we rephrased these statements.

Now, the usage environment contains only the linear variables in the
body of the binder for the variable a usage environment is attached to.