Realize the fragmented usage environment idea does not work trivially
It's not easy at all to think about the semantic linearity in the lazy
language, but it looks like I'm able to prove that linear double free is
well-typed, which is awful.

In restrospect, the branch cases cannot have the things that are used in
the scrutinee at all, and likely need to consume the bound things too.

We just need to realise why exactly those intermediate programs in the
reverse binder swap preserve linearity sematnically, and then how we
can make them well-typed