Add some horrendous hacks to make things work.
Currently, we can parse some things but a lot of
typechecking types is broken when we shrunk down Core to our Linear Core
that only has multiplicity abstractions and little else.

For example, unification of type variables was hacked around.
Type synonyms and Implicit Parameters were hacked around.

Now we're bumping against Coercions, since we can't match IO with
(State# -> (#,#)) (which is already a dumbed version of something like
(State# -> (# (), State# #)).

I think the right thing to do from here on is to completely disable
typechecking types -- the linter already does that.

/We should only care about linting linearity, and leave types checking to
the non-linear linter./