No, the e' in the Sigma is not substituted
There are two cases:

In the "e is in WHNF" we don't need that induction hypothesis to deduce
the result

In the "e is not in WHNF", the resources introduced in e' by the case
are "irrelevant", so we can move the \D bound variables to linear and
wipe that from the environment to reconstruct (e, e') and apply the
induction hypothesis. We also need to do a case analysis on whether z
occurs or not in e' for some details. But it all works out.